update login/signup and profile editing flow, and modify mentor/mentee models (#35)
trang updated the models so that the sign up page hosts a binary flag for mentor/mentee for the user to select. this allows us to drop the "do you want to be a mentor / mentee" page, and takes the user to fill out their profile. mentor sign up approval will go in the backlog